# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM Nekretnine API",
        "description": "API dokumentacija za CRM sistem nekretnina",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/kupci": {
            "get": {
                "tags": [
                    "Kupci"
                ],
                "summary": "Lista kupaca",
                "operationId": "1972846269edf52a0953501fb90ca9e4",
                "responses": {
                    "200": {
                        "description": "Lista kupaca",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Kupac"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Kupci"
                ],
                "summary": "Kreiranje kupca",
                "operationId": "cc2c641b357454ba6e866fdbbc21bc5e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ime",
                                    "prezime"
                                ],
                                "properties": {
                                    "ime": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Ana"
                                    },
                                    "prezime": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Anić"
                                    },
                                    "telefon": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "+381641112223",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "ana@email.com",
                                        "nullable": true
                                    },
                                    "budzet": {
                                        "type": "number",
                                        "example": 80000,
                                        "nullable": true
                                    },
                                    "lokacija": {
                                        "type": "string",
                                        "maxLength": 120,
                                        "example": "Beograd",
                                        "nullable": true
                                    },
                                    "napomena": {
                                        "type": "string",
                                        "example": "Traži 2-soban stan, blizu centra",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kupac kreiran",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Kupac"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
        "/api/kupci/{id}": {
            "get": {
                "tags": [
                    "Kupci"
                ],
                "summary": "Detalji kupca",
                "operationId": "f5927e5da62a0c7cc3fcd568fc985431",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kupac",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Kupac"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kupac nije pronađen"
                    }
                }
            },
            "put": {
                "tags": [
                    "Kupci"
                ],
                "summary": "Izmena kupca",
                "operationId": "1c2aa0d7289bebe9b272a7c20a076fae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ime": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Ana",
                                        "nullable": true
                                    },
                                    "prezime": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Anić",
                                        "nullable": true
                                    },
                                    "telefon": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "+381641112223",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255,
                                        "example": "ana@email.com",
                                        "nullable": true
                                    },
                                    "budzet": {
                                        "type": "number",
                                        "example": 90000,
                                        "nullable": true
                                    },
                                    "lokacija": {
                                        "type": "string",
                                        "maxLength": 120,
                                        "example": "Novi Sad",
                                        "nullable": true
                                    },
                                    "napomena": {
                                        "type": "string",
                                        "example": "Prioritet: parking",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kupac izmenjen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Kupac"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kupac nije pronađen"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Kupci"
                ],
                "summary": "Brisanje kupca",
                "operationId": "51890156eef357f5099a15877d9d824d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kupac obrisan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kupac obrisan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kupac nije pronađen"
                    }
                }
            }
        },
        "/api/kupci/search": {
            "get": {
                "tags": [
                    "Kupci"
                ],
                "summary": "Pretraga kupaca (paginacija + sortiranje)",
                "description": "Primer: /api/kupci/search?q=...&page=1&per_page=10&sort_by=prezime&sort_dir=asc",
                "operationId": "d86310dddc980975e8b2c99a822acd87",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 100
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        },
                        "example": 10
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ime",
                                "prezime",
                                "email",
                                "telefon",
                                "budzet",
                                "lokacija",
                                "created_at"
                            ]
                        },
                        "example": "prezime"
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        },
                        "example": "asc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginirani rezultati",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Kupac"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Kupac": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ime": {
                        "type": "string",
                        "example": "Ana"
                    },
                    "prezime": {
                        "type": "string",
                        "example": "Anić"
                    },
                    "telefon": {
                        "type": "string",
                        "example": "+381641112223",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ana@email.com",
                        "nullable": true
                    },
                    "budzet": {
                        "type": "number",
                        "example": 80000,
                        "nullable": true
                    },
                    "lokacija": {
                        "type": "string",
                        "example": "Beograd",
                        "nullable": true
                    },
                    "napomena": {
                        "type": "string",
                        "example": "Traži 2-soban stan",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}